PSF: address SonarCloud issues in PR #2542
- pointstarrenderer: use init-statement for delta; drop stale comment
- render: cast gl::maxPointSize to float
- psfstarvertexbuffer: use std::array and std::vector instead of C arrays
- configfile: group star rendering fields into a nested struct

+++ src/celengine/render.cpp
@@ -3718,7 +3718,7 @@ void Renderer::renderPointStars(const StarDatabase& starDB,
     starRenderer.maxIrradiance     = starMaxIrradiance;
     starRenderer.exposure          = starExposure;
     starRenderer.pointScale        = static_cast<float>(screenDpi) / 96.0f;
-    starRenderer.maxPointSize      = celestia::gl::maxPointSize;
+    starRenderer.maxPointSize      = static_cast<float>(celestia::gl::maxPointSize);
     starRenderer.largeGlowStars.clear();
 
     // = 1.0 at startup
